Added material-ui integration. Installed material ui, together with roboto fonts and material icons. Restructured files. Moved App component into components folder. Changed js files to jsx.
