+++ src/components/App.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import axios from 'axios';
-import baseUrl from './config';
+import baseUrl from '../config';
 
 const App = () => {
