+++ src/components/__tests__/App.test.jsx
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react';
 import React from 'react';
-import App from './App';
+import App from '../App';
 
 describe('App component', () => {
     it('should render app with error', () => {
