+++ src/index.jsx
@@ -1,7 +1,12 @@
 import React from 'react';
+// import ReactDOM from 'react-dom';
 import ReactDOM from "react-dom/client";
-// import ReactDOM from 'react-dom/client';
-import App from './App';
+import App from './components/App';
+import '@fontsource/roboto/300.css';
+import '@fontsource/roboto/400.css';
+import '@fontsource/roboto/500.css';
+import '@fontsource/roboto/700.css';
+
 
 // ReactDOM.render(<App />, document.getElementById("root"));
